Remove 'nowhere' from Both sides N
In Plover's newest `main.json`, the brief for `nowhere` is `TPHO*ER` and not `TPH-R` Because this brief contains `OE` which isn't introduced at this point in the textbook I think it would be in the best interest to remove `nowhere`. I did my best to come up with a replacement sentence that continues the "plot" with the third outline finding practice.