Student has not learned how to chord the word "your" yet
In this exercise, the student is tasked to write the sentence, "I covet your oven. I want to have it and when I do I will warm my rat with it." But the learner has on knowledge about beginning y sounds, and in no way could figure out on his own that it was "WROUR." I changed the sentence so that it is possible to write with the student's current knowledge.